Interface Overhaul + Qt6 Port
Main point of these changes, besides the need to port to Qt6, is to give access to the most used options in a single page that appears first when the interface is opened.

The changes:
Moved the shift mode selection to the first page
Removed the Mode tab and page entirely
Toggles have been also moved to the first page
Renamed the tab (page) the Overview tab
Renamed the settings tab to Keyboard tab (since the only Options left are keyboard related)
Moved EC version and build date from the overview page to the about page
 
More changes:
The interface won't be resizable anymore, that's for avoiding it getting tiled by tiling window managers or scripts that emulate tiling on normal (stacking) desktop environments. this was done to avoid wasted space in case the interface was the only window visible on the screen.

Side changes:
About page: mention msi-ec as the driver and add a github link for it.
Debug page: add warning against randomly writing to EC memory.

The new interface was tested with multiple Qt themes and seems to look and function fine.